♻️ Light refactoring
