--- src/fixLayer.js
@@ -1,8 +1,0 @@
-export default {
-  id: 'fixPoints',
-  type: 'circle',
-  paint: {
-    'circle-radius': 4,
-    'circle-color': '#fff',
-  },
-};
